requirements.txt
abstract_singleton==1.0.1
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
anyio==3.7.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
auto_gpt_plugin_template==0.0.3
autoflake==2.1.1
black==23.3.0
blinker==1.6.2
blis==0.7.9
CacheControl==0.13.1
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
coverage==7.2.5
cryptography==40.0.0
cssselect==1.2.0
cymem==2.0.7
dataclasses-json==0.5.8
diskcache==5.6.1
distlib==0.3.6
distro==1.8.0
dnspython==2.3.0
docker==6.1.1
duckduckgo-search==2.9.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
exceptiongroup==1.1.1
faiss-cpu==1.7.4
filelock==3.12.2
firebase-admin==6.1.0
flake8==6.0.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.10
frozenlist==1.3.3
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
gTTS==2.3.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.0
httplib2==0.22.0
httpx==0.24.0
identify==2.5.24
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.198
langchainplus-sdk==0.0.8
langcodes==3.3.0
loguru==0.7.0
lxml==4.9.2
Markdown==3.3.7
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.4.3
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nodeenv==1.7.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-python-client==0.13.4
openapi-schema-pydantic==1.2.4
orjson==3.8.10
outcome==1.2.0
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
pathy==0.10.1
pickle-mixin==1.0.2
pinecone-client==2.2.1
platformdirs==3.5.3
playsound==1.2.2
pluggy==1.0.0
pre-commit==3.3.1
preshed==3.0.8
proto-plus==1.22.2
protobuf==4.23.2
py-cpuinfo==9.0.0
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.9
pyflakes==3.0.1
PyJWT==2.7.0
pymdown-extensions==9.11
pyOpenSSL==23.1.1
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.3.2
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
pytest-recording==0.12.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
pyyaml_env_tag==0.1
readability-lxml==0.8.1
redis==4.5.5
regex==2023.6.3
requests==2.31.0
rsa==4.9
selenium==4.1.4
shellingham==1.5.0.post1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.16
srsly==2.4.6
tenacity==8.2.2
thinc==8.1.10
tiktoken==0.4.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
urllib3-secure-extra==0.1.0
vcrpy==4.2.1
virtualenv==20.23.0
wasabi==1.1.2
watchdog==3.0.0
webdriver-manager==3.8.6
websocket-client==1.5.3
Werkzeug==2.3.6
win32-setctime==1.1.0
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2